--- a/hw10_program_optimization/stats.go
+++ b/hw10_program_optimization/stats.go
@@ -1,66 +1,57 @@
 package hw10programoptimization
 
 import (
-	"encoding/json"
+	"bufio"
+	"errors"
 	"fmt"
 	"io"
-	"regexp"
+	"net/mail"
 	"strings"
+
+	jsoniter "github.com/json-iterator/go"
 )
 
 type User struct {
-	ID       int
-	Name     string
-	Username string
-	Email    string
-	Phone    string
-	Password string
-	Address  string
+	Email string
 }
-
 type DomainStat map[string]int
 
+var (
+	ErrUnmarshalData      = errors.New("unmarshal data error")
+	ErrInvalidEmail       = errors.New("invalid email error")
+	ErrInappropriateEmail = errors.New("inappropriate email error")
+	json                  = jsoniter.ConfigCompatibleWithStandardLibrary
+)
+
 func GetDomainStat(r io.Reader, domain string) (DomainStat, error) {
-	u, err := getUsers(r)
-	if err != nil {
-		return nil, fmt.Errorf("get users error: %w", err)
-	}
-	return countDomains(u, domain)
-}
-
-type users [100_000]User
-
-func getUsers(r io.Reader) (result users, err error) {
-	content, err := io.ReadAll(r)
-	if err != nil {
-		return
-	}
-
-	lines := strings.Split(string(content), "\n")
-	for i, line := range lines {
-		var user User
-		if err = json.Unmarshal([]byte(line), &user); err != nil {
-			return
-		}
-		result[i] = user
-	}
-	return
-}
-
-func countDomains(u users, domain string) (DomainStat, error) {
+	scanner := bufio.NewScanner(r)
 	result := make(DomainStat)
-
-	for _, user := range u {
-		matched, err := regexp.Match("\\."+domain, []byte(user.Email))
-		if err != nil {
-			return nil, err
+	user := &User{}
+	domainSuffix := "." + domain
+	for scanner.Scan() {
+		if err := json.Unmarshal(scanner.Bytes(), user); err != nil {
+			return nil, fmt.Errorf("%s: %w", string(scanner.Bytes()), ErrUnmarshalData)
 		}
 
-		if matched {
-			num := result[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])]
-			num++
-			result[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])] = num
+		if _, err := checkEmail(user.Email, domainSuffix); err != nil {
+			continue
 		}
+
+		emailDomain := strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])
+		result[emailDomain]++
 	}
+
 	return result, nil
+}
+
+func checkEmail(email string, domainSuffix string) (bool, error) {
+	if !strings.HasSuffix(email, domainSuffix) {
+		return false, ErrInappropriateEmail
+	}
+
+	if _, err := mail.ParseAddress(email); err != nil {
+		return false, ErrInvalidEmail
+	}
+
+	return true, nil
 }