--- a/hw10_program_optimization/stats_test.go
+++ b/hw10_program_optimization/stats_test.go
@@ -1,3 +1,4 @@
+//go:build !bench
 // +build !bench
 
 package hw10programoptimization
@@ -35,5 +36,26 @@ func TestGetDomainStat(t *testing.T) {
 		result, err := GetDomainStat(bytes.NewBufferString(data), "unknown")
 		require.NoError(t, err)
 		require.Equal(t, DomainStat{}, result)
+	})
+
+	malformedData := `{"Id":1,"Name":"Howard Mendoza","Username":"0Oliver","Email":"aliquid_qui_ea@Browsedrive.gov","Phone":"6-866-899-36-79","Password":"InAQJvsq`
+	t.Run("unmarshal data error", func(t *testing.T) {
+		result, err := GetDomainStat(bytes.NewBufferString(malformedData), "com")
+		require.Error(t, err)
+		require.Nil(t, result)
+	})
+
+	InvalidEmailData := `{"Id":1,"Name":"Howard Mendoza","Username":"0Oliver","Email":"5Moore.com@Teklist.com","Phone":"6-866-899-36-79","Password":"InAQJvsq","Address":"Blackbird Place 25"}
+{"Id":2,"Name":"Jesse Vasquez","Username":"qRichardson","Email":"Teklist.com@5Moore.com","Phone":"9-373-949-64-00","Password":"SiZLeNSGn","Address":"Fulton Hill 80"}
+{"Id":3,"Name":"Jesse Vasquez","Username":"qRichardson","Email":"RoseSmithBrowsecat.com","Phone":"9-373-949-64-00","Password":"SiZLeNSGn","Address":"Fulton Hill 80"}
+{"Id":4,"Name":"Clarence Olson","Username":"RachelAdams","Email":"RoseSmith@Browsecat.com","Phone":"988-48-97","Password":"71kuz3gA5w","Address":"Monterey Park 39"}`
+	t.Run("invalid email error", func(t *testing.T) {
+		result, err := GetDomainStat(bytes.NewBufferString(InvalidEmailData), "com")
+		require.NoError(t, err)
+		require.Equal(t, DomainStat{
+			"teklist.com":   1,
+			"5moore.com":    1,
+			"browsecat.com": 1,
+		}, result)
 	})
 }