## Домашнее задание №10 «Оптимизация программы»

### Чек-лист студента ([Что это?](https://github.com/OtusGolang/home_work/wiki/%D0%9A%D0%BE%D0%BC%D0%BC%D0%B5%D0%BD%D1%82%D0%B0%D1%80%D0%B8%D0%B8-%D0%BA-%D1%87%D0%B5%D0%BA-%D0%BB%D0%B8%D1%81%D1%82%D1%83-%D1%81%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0))
- [x] Я перечитал задание после решения. [Зачем перечитывать задание?](https://github.com/OtusGolang/home_work/wiki/%D0%9A%D0%BE%D0%BC%D0%BC%D0%B5%D0%BD%D1%82%D0%B0%D1%80%D0%B8%D0%B8-%D0%BA-%D1%87%D0%B5%D0%BA-%D0%BB%D0%B8%D1%81%D1%82%D1%83-%D1%81%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0#user-content-%D0%97%D0%B0%D1%87%D0%B5%D0%BC-%D0%BF%D0%B5%D1%80%D0%B5%D1%87%D0%B8%D1%82%D1%8B%D0%B2%D0%B0%D1%82%D1%8C-%D0%B7%D0%B0%D0%B4%D0%B0%D0%BD%D0%B8%D0%B5)
- [x] Я запустил `go mod tidy`.
- [x] Я удалил `.sync` файл. [Зачем его удалять?](https://github.com/OtusGolang/home_work/wiki/%5B%D0%A1%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0%D0%BC%5D-%D0%9F%D1%80%D0%BE%D1%86%D0%B5%D1%81%D1%81-%D1%81%D0%B4%D0%B0%D1%87%D0%B8-%D0%94%D0%97#user-content-%D0%92%D0%B0%D1%80%D0%B8%D0%B0%D0%BD%D1%82-2)
- [x] Я написал бенчмарк и приложил результаты benchstat.

### Критерии оценки
- [ ] Пайплайн зелёный и нет попытки «обмануть» систему - 4 балла
- [ ] Добавлены юнит-тесты - до 3 баллов
- [ ] Понятность и чистота кода - до 3 баллов

#### Зачёт от 7 баллов

**До правок:** 
```
goos: linux
goarch: amd64
pkg: github.com/Roman-Otus-Learning/roman-otus-go-2024-01/hw10_program_optimization
cpu: 13th Gen Intel(R) Core(TM) i5-13600KF
BenchmarkGetDomainStat-20    	      14	  73306765 ns/op	136045896 B/op	 1700089 allocs/op
BenchmarkGetDomainStat-20    	      15	  72660910 ns/op	136045901 B/op	 1700089 allocs/op
BenchmarkGetDomainStat-20    	      15	  72844184 ns/op	136045937 B/op	 1700089 allocs/op
BenchmarkGetDomainStat-20    	      15	  72602315 ns/op	136045899 B/op	 1700089 allocs/op
BenchmarkGetDomainStat-20    	      14	  72961902 ns/op	136045937 B/op	 1700089 allocs/op
BenchmarkGetDomainStat-20    	      14	  72554590 ns/op	136045932 B/op	 1700089 allocs/op
BenchmarkGetDomainStat-20    	      15	  72871443 ns/op	136045902 B/op	 1700089 allocs/op
BenchmarkGetDomainStat-20    	      15	  72519735 ns/op	136046094 B/op	 1700089 allocs/op
BenchmarkGetDomainStat-20    	      15	  72696923 ns/op	136045884 B/op	 1700089 allocs/op
BenchmarkGetDomainStat-20    	      14	  72769158 ns/op	136045955 B/op	 1700089 allocs/op
PASS
```

**После правок:** 
```
goos: linux
goarch: amd64
pkg: github.com/Roman-Otus-Learning/roman-otus-go-2024-01/hw10_program_optimization
cpu: 13th Gen Intel(R) Core(TM) i5-13600KF
BenchmarkGetDomainStat-20    	  370467	      3174 ns/op	    5308 B/op	      80 allocs/op
BenchmarkGetDomainStat-20    	  349424	      3209 ns/op	    5308 B/op	      80 allocs/op
BenchmarkGetDomainStat-20    	  372508	      3287 ns/op	    5308 B/op	      80 allocs/op
BenchmarkGetDomainStat-20    	  352129	      3219 ns/op	    5308 B/op	      80 allocs/op
BenchmarkGetDomainStat-20    	  373663	      3211 ns/op	    5308 B/op	      80 allocs/op
BenchmarkGetDomainStat-20    	  369702	      3181 ns/op	    5308 B/op	      80 allocs/op
BenchmarkGetDomainStat-20    	  368740	      3179 ns/op	    5308 B/op	      80 allocs/op
BenchmarkGetDomainStat-20    	  370053	      3184 ns/op	    5308 B/op	      80 allocs/op
BenchmarkGetDomainStat-20    	  367141	      3193 ns/op	    5308 B/op	      80 allocs/op
BenchmarkGetDomainStat-20    	  364708	      3183 ns/op	    5308 B/op	      80 allocs/op
```

**После правок CR:** 
```
BenchmarkGetDomainStat-20         224413              5202 ns/op            6076 B/op        128 allocs/op
BenchmarkGetDomainStat-20         223364              5181 ns/op            6076 B/op        128 allocs/op
BenchmarkGetDomainStat-20         232375              5106 ns/op            6076 B/op        128 allocs/op
BenchmarkGetDomainStat-20         229051              5105 ns/op            6076 B/op        128 allocs/op
BenchmarkGetDomainStat-20         228760              5113 ns/op            6076 B/op        128 allocs/op
BenchmarkGetDomainStat-20         232446              5104 ns/op            6076 B/op        128 allocs/op
BenchmarkGetDomainStat-20         230742              5110 ns/op            6076 B/op        128 allocs/op
BenchmarkGetDomainStat-20         231355              5105 ns/op            6076 B/op        128 allocs/op
BenchmarkGetDomainStat-20         230648              5119 ns/op            6076 B/op        128 allocs/op
BenchmarkGetDomainStat-20         230328              5103 ns/op            6076 B/op        128 allocs/op
PASS

```

**benchstat:** 
```
goos: linux
goarch: amd64
pkg: github.com/Roman-Otus-Learning/roman-otus-go-2024-01/hw10_program_optimization
cpu: 13th Gen Intel(R) Core(TM) i5-13600KF
                 │    old2.txt     │               new2.txt               │
                 │     sec/op      │   sec/op     vs base                 │
GetDomainStat-20   72733.041µ ± 0%   3.189µ ± 1%  -100.00% (p=0.000 n=10)

                 │     old2.txt      │               new2.txt                │
                 │       B/op        │     B/op      vs base                 │
GetDomainStat-20   132857.341Ki ± 0%   5.184Ki ± 0%  -100.00% (p=0.000 n=10)

                 │    old2.txt     │              new2.txt               │
                 │    allocs/op    │ allocs/op   vs base                 │
GetDomainStat-20   1700089.00 ± 0%   80.00 ± 0%  -100.00% (p=0.000 n=10)
```